# Conflicts:
#	src/main/java/de/qStivi/listeners/ReactionRoles.java

+++ src/main/java/de/qStivi/Card.java
@@ -3,7 +3,7 @@ package de.qStivi;
 public class Card {
 
     public Suit suit;
-    public int value;
+    public double value;
     public String emote;
 
     public Card(Suit suit, int value, String emote) {
